Change forecast_data table accept all Metrics
