+++ packages/api/src/wind-wave-data/wind-wave-data.entity.ts
@@ -12,14 +12,6 @@ import {
 import { Site } from '../sites/sites.entity';
 import { Metric } from '../time-series/metrics.entity';
 
-export enum WindWaveMetric {
-  SIGNIFICANT_WAVE_HEIGHT = Metric.SIGNIFICANT_WAVE_HEIGHT,
-  WAVE_MEAN_DIRECTION = Metric.WAVE_MEAN_DIRECTION,
-  WAVE_MEAN_PERIOD = Metric.WAVE_MEAN_PERIOD,
-  WIND_SPEED = Metric.WIND_SPEED,
-  WIND_DIRECTION = Metric.WIND_DIRECTION,
-}
-
 @Entity()
 @Unique('one_row_per_site_per_metric', ['site', 'metric'])
 @Index(['site'])
@@ -37,8 +29,8 @@ export class ForecastData {
   @Column({ type: 'float', nullable: false })
   value: number;
 
-  @Column({ type: 'enum', enum: WindWaveMetric, nullable: false })
-  metric: WindWaveMetric;
+  @Column({ type: 'enum', enum: Metric, nullable: false })
+  metric: Metric;
 
   @CreateDateColumn()
   createdAt: Date;
